+++ miniscope_io/devices/__init__.py
@@ -1,3 +1,8 @@
 """
-Control interfaces for external hardware devices
+Devices!
 """
+
+from miniscope_io.devices.camera import Camera, Miniscope
+from miniscope_io.devices.device import Device
+
+__all__ = ["Camera", "Device", "Miniscope"]