+++ miniscope_io/models/__init__.py
@@ -2,10 +2,30 @@
 Data models :)
 """
 
-from miniscope_io.models.models import Container, MiniscopeConfig, MiniscopeIOModel
+from miniscope_io.models.models import (
+    Container,
+    MiniscopeConfig,
+    MiniscopeIOModel,
+    PipelineModel,
+)
+from miniscope_io.models.pipeline import (
+    Node,
+    Pipeline,
+    PipelineConfig,
+    Sink,
+    Source,
+    Transform,
+)
 
 __all__ = [
     "Container",
     "MiniscopeConfig",
     "MiniscopeIOModel",
+    "Node",
+    "Pipeline",
+    "PipelineConfig",
+    "PipelineModel",
+    "Transform",
+    "Sink",
+    "Source",
 ]