+++ miniscope_io/sources/__init__.py
@@ -1,0 +1,7 @@
+"""
+Control interfaces for external hardware sources
+"""
+
+from miniscope_io.sources.opalkelly import okDev
+
+__all__ = ["okDev"]